PR feedbacks
Removing ProductVersions enum, replacing by a default "None" value for the product_version parameter. If it's none, the used version will be the latest installed one (as long as it's >= V23.2) . If a specific version's value is defined, we will just check that it is installed. If so, we will launch that specific version.