Expose the interference tool in python
The interference tool was exposed in the API server but not in python. One thing to notice is that this tool doesn't do tracking so the fix method will fail with the current API server. A pending PR is adding a try catch to allow lack of tracking to throw an exception.