Exposure of detect llogo implementation
Exposure of detect logo implementation.  Logo is defined as a problem area described by a list of faces.  It is not a list of logs because the detection provides a list of faces without any distinction (other than per body).  The problem area can be removed with a fix command or a find and remove can remove them all